requirements.txt
jupyter=1.0.0
matplotlib=3.1.3
notebook=6.0.3
numpy=1.18.1
pandas=1.0.3
pytorch=1.5.0
scikit-learn=0.22.1
scipy=1.4.1
seaborn=0.10.1
tqdm=4.46.0
